abstract logic into a custom hook
